fix: Footer coming in middle of the page when no content
After removing the position fixed of the footer in the last commit,
I realized after some testing that the reason it was put there was to not let it come in the middle of the webpage when there is not content.
Now after its removal, the issue came up, so I made the Background  a flexbox and set the right properties of it and its children so that the footer stays at the bottom even if there isn't any content (the puzzles).